chore(crdt-dash): break components into files
