+++ src/components/social-media-graphics/race/social-card.js
@@ -3,17 +3,16 @@ import classnames from 'classnames'
 
 import { renderedComponent } from '~plugins/gatsby-render-components'
 
-import { FormatNumber } from '~components/utils/format'
-
 import Logo from '~images/ctp-icon-small.png'
 import CarLogo from '~images/car-logo-small.png'
 
 import SocialCardFootnotes from './footnotes'
 import SocialCardHeader from './header'
 import NoDataSocialCard from './no-data'
-import { getStateStatus, getGroups, getBarWidth } from './utils'
+import { getStateStatus, getGroups } from './utils'
 
 import socialCardStyle from './social-card.module.scss'
+import ChartRow from './chart-row'
 
 const getFootnoteStatuses = (stateGroups, stateName) => {
   const groups = stateGroups
@@ -52,123 +51,6 @@ const getFootnoteStatuses = (stateGroups, stateName) => {
     showSmallNFootnote,
     asteriskFootnote,
   }
-}
-
-const ChartBar = ({
-  style,
-  metric,
-  worstMetricValue,
-  square,
-  isOneChart,
-  showAsterisk = false,
-  isDeaths,
-}) => {
-  // gets the width of the bar for the bar charts
-  const widthPercentage = (metric / worstMetricValue) * 100
-
-  const BarContent = ({ value, asterisk = false }) => (
-    <span>
-      <FormatNumber number={value} />
-      {asterisk && '*'}
-    </span>
-  )
-
-  return (
-    <div className={socialCardStyle.barContainer}>
-      {isDeaths && <div className={socialCardStyle.deathBarSpacer} />}
-      <div
-        className={classnames(
-          socialCardStyle.bar,
-          style,
-          widthPercentage !== 0 && socialCardStyle.hasInnerLabel,
-        )}
-        style={{
-          width: `${getBarWidth(
-            metric,
-            worstMetricValue,
-            square,
-            isOneChart,
-          )}px`,
-        }}
-      >
-        {widthPercentage > 50 && (
-          <BarContent value={metric} asterisk={showAsterisk} />
-        )}
-      </div>
-      {widthPercentage <= 50 && (
-        <BarContent value={metric} asterisk={showAsterisk} />
-      )}
-    </div>
-  )
-}
-
-const ChartRow = ({
-  group,
-  stateStatus,
-  worstCasesValue,
-  worstDeathsValue,
-  square,
-}) => {
-  const { label, style, cases, deaths, showAsterisk, showCross } = group
-
-  const BarLabel = ({ barLabel, cross }) => (
-    <span className={socialCardStyle.barLabel}>
-      {barLabel}
-      {cross && '†'}
-    </span>
-  )
-
-  const InsufficientDataPlaceholder = ({ isDeaths }) => {
-    const nullValue = 'No data reported' // the value to show for the empty state
-    return (
-      <span
-        className={classnames(
-          socialCardStyle.insufficientData,
-          isDeaths && socialCardStyle.insufficientDataDeaths,
-        )}
-      >
-        {nullValue}
-      </span>
-    )
-  }
-
-  return (
-    <>
-      <BarLabel barLabel={label} cross={showCross} />
-      {!stateStatus.deathsOnly && (
-        <>
-          {cases === null ? (
-            <InsufficientDataPlaceholder />
-          ) : (
-            <ChartBar
-              style={style}
-              metric={cases}
-              worstMetricValue={worstCasesValue}
-              square={square}
-              isOneChart={stateStatus.oneChart}
-            />
-          )}
-        </>
-      )}
-      {!stateStatus.casesOnly && (
-        <>
-          {deaths === null ? (
-            <InsufficientDataPlaceholder isDeaths />
-          ) : (
-            <ChartBar
-              style={style}
-              metric={deaths}
-              worstMetricValue={worstDeathsValue}
-              square={square}
-              showAsterisk={showAsterisk}
-              isOneChart={stateStatus.oneChart}
-              isDeaths
-            />
-          )}
-        </>
-      )}
-    </>
-  )
 }
 
 const StateRaceSocialCard = renderedComponent(
